+++ webapi/config.py
@@ -1,0 +1,3 @@
+DEBUG = True # Turns on debugging features in Flask
+TESTING = True
+DATASTORE = "/Users/gmaze/git/github/euroargodev/VirtualFleet_recovery/webapi/myapp/static"
