+++ webapi/myapp/docs/predict.yml
