+++ webapi/myapp/static/js/trigger.js
