+++ webapi/myapp/templates/list2.html
@@ -172,6 +172,18 @@
                 N floats: {{NFLOATS}}
               </a>
             </li>
+            <li class="nav-item">
+              <a class="nav-link disabled" aria-current="page" href="#">
+                <span data-feather="activity" class="align-text-bottom"></span>
+                Parking Depth: {{CFG_PARKING_DEPTH}}
+              </a>
+            </li>
+            <li class="nav-item">
+              <a class="nav-link disabled" aria-current="page" href="#">
+                <span data-feather="clock" class="align-text-bottom"></span>
+                Cycle Duration: {{CFG_CYCLE_DURATION}}
+              </a>
+            </li>
           </ul>
 
           <h6 class="sidebar-heading d-flex justify-content-between align-items-center px-3 mt-4 mb-1 text-muted text-uppercase">
