+++ webapi/myapp/utils/__init__.py
@@ -1,0 +1,11 @@
+#!/bin/env python
+# -*coding: UTF-8 -*-
+#
+# HELP
+#
+# Created by gmaze on 02/12/2022
+__author__ = 'gmaze@ifremer.fr'
+
+import os
+import sys
+import xarray as xr
